+++ packages/opentelemetry/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @prairielearn/opentelemetry
 
+## 2.1.9
+
+### Patch Changes
+
+- 0e1aa6e: Upgrade all Sentry and OpenTelemetry packages
+- 4a8b376: Upgrade all JavaScript dependencies
+
 ## 2.1.8
 
 ### Patch Changes